==> analysis/main/02_xor-some_prereg_plots_final.Rmd <==
---
title: 'Xor-Some Preregistered Study: Visualizations'
author: "Polina Tsvilodub"
date: "10/18/2021"
output: github_document
---

```{r setup, include=FALSE}
library(tidyverse)
library(tidyboot)
```

Read data created in preprocessing script:
```{r, warning=FALSE, message=FALSE, results='hide'}
d_critical_zScored_wide <- read_csv("./../../data/main/results_prereg_tidy_final_zScored_wide.csv")
d_critical_zScored <- read_csv("./../../data/main/results_prereg_tidy_final_zScored_long.csv")
d_test <- read_csv("./../../data/main/results_prereg_raw_final_test.csv")
d_critical_zScored_wide_xor_priors_order <- read_csv("./../../data/main/results_prereg_tidy_final_zScored_wide_xor_priors-order.csv")
d_critical_zScored_wide_xor_priors_type <- read_csv("./../../data/main/results_prereg_tidy_final_zScored_wide_xor_priors-type.csv")

d_critical_zScored_meanFirst_noDupl_zScored_wide <- 
  read_csv("./../../data/main/results_prereg_tidy_noDuplicates_zScored_meansFirst_wide.csv")

```

## Plots

Comprehension question performance:
```{r}
d_test_ci <- d_test %>% group_by(test_condition) %>% 
  tidyboot_mean(column = response) 

d_test %>% 
  ggplot(., aes(x = test_condition, y = response)) +
  geom_point(size = 2, alpha = 0.3, position = position_jitter(width = 0.1)) +
  geom_point(data = d_test_ci, aes(x = test_condition, y = mean), color = "red", 
             size = 4) +
  facet_wrap(~main_type)
```

Plot predictor ratings against inference strength ratings by trigger:
```{r, echo=FALSE, message=FALSE, warning=FALSE}
d_critical_zScored_wide %>%
  ggplot(., aes(x = relevance, y = target)) +
  geom_point(size = 2, alpha = 0.7) +
  ylab("Inference strength ratings") +
  geom_smooth(method = "lm") +
  facet_wrap(~main_type, ncol = 1) -> p.rel.z

d_critical_zScored_wide %>%
  ggplot(., aes(x = competence, y = target)) +
  geom_point(size = 2, alpha = 0.7) +
  geom_smooth(method = "lm") +
  ylab("") +
  facet_wrap(~main_type, ncol = 1) -> p.comp.z

d_critical_zScored_wide %>%
  ggplot(., aes(x = prior, y = target)) +
  geom_point(size = 2, alpha = 0.7) +
  geom_smooth(method = "lm") +
  ylab("") +
  facet_wrap(~main_type, ncol = 1) -> p.pri.z

gridExtra::grid.arrange(p.rel.z, p.comp.z, p.pri.z, ncol = 3) 
```
Replicate plot above on data where means were computed over the prior ratings for or before z-scoring:
```{r}
d_critical_zScored_meanFirst_noDupl_zScored_wide %>%
  ggplot(., aes(x = relevance, y = target)) +
  geom_point(size = 2, alpha = 0.7) +
  ylab("Inference strength ratings") +
  geom_smooth(method = "lm") +
  facet_wrap(~main_type, ncol = 1) -> p.rel.z

d_critical_zScored_meanFirst_noDupl_zScored_wide %>%
  ggplot(., aes(x = competence, y = target)) +
  geom_point(size = 2, alpha = 0.7) +
  geom_smooth(method = "lm") +
  ylab("") +
  facet_wrap(~main_type, ncol = 1) -> p.comp.z

d_critical_zScored_meanFirst_noDupl_zScored_wide %>%
  ggplot(., aes(x = prior, y = target)) +
  geom_point(size = 2, alpha = 0.7) +
  geom_smooth(method = "lm") +
  ylab("") +
  facet_wrap(~main_type, ncol = 1) -> p.pri.z

gridExtra::grid.arrange(p.rel.z, p.comp.z, p.pri.z, ncol = 3) 
```

```{r}
d_critical_zScored %>% 
  filter(block == "xor" | block == "some") %>%
  ggplot(., aes(x = prior_class, y = response_centered)) +
  geom_point(size = 2, alpha = 0.5) +
  geom_smooth(method="lm") +
  ylab("Inference strength ratings") +
  facet_wrap(block~class_condition) +
  xlab("Anticipated categorization of the items") +
  ggtitle("Inference strength ratings by-predictor")
  
```

```{r, fig.height=30, fig.width=10}
# plot by-item ratings vs our prior categorizations
d_critical_long_grandMeans <- d_critical_zScored %>% 
  mutate(title = paste(title, exp_condition, sep = "_"),
         block = ifelse(block == "xor" | block == "some", "target", block)) %>%
  filter(block == class_condition | (block == "target" & class_condition == "relevance")) %>%
  group_by(main_type, block) %>% 
  summarize(mean_resp = mean(response))

d_critical_long_expectedCats <- d_critical_zScored %>% 
  mutate(title = paste(title, exp_condition, sep = "_"),
         block = ifelse(block == "xor" | block == "some", "target", block)) %>%
  filter(block == class_condition | (block == "target" & class_condition == "relevance")) %>%
  group_by(main_type, block, title) %>% 
  summarize(mean_resp_block = mean(response))

d_critical_zScored %>% 
  mutate(title = paste(title, exp_condition, sep = "_"),
         block = ifelse(block == "xor" | block == "some", "target", block)) %>%
  left_join(., d_critical_long_grandMeans, by = c("main_type", "block")) %>% 
  filter(block == class_condition | ((block == "target") & (class_condition == "relevance"))) -> d_long_plot_data

left_join(d_long_plot_data, d_critical_long_expectedCats, by = c("main_type", "block", "title")) %>% 
  mutate(expected_block = ifelse(block != "target", prior_class * 100, NA),
         block = factor(block, levels = c("relevance", "competence", "prior", "target"))) -> d_long_plot_full

d_long_plot_full %>%
  ggplot(., aes(x = block, y = response, color = block)) +
  geom_point(alpha = 0.7, position = position_jitter(width = 0.1)) +
  geom_hline(aes(yintercept = mean_resp), alpha = 0.7) +
  geom_point(aes(y = mean_resp), color = "black", shape = 2) +
  geom_point(aes(y = mean_resp_block, fill = block), color = "black") +
  geom_point(aes(y = expected_block), color="red", size = 2.5, na.rm = T) +
  ylab("Raw responses to respective questions") +
  facet_wrap(main_type~title, ncol = 4) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  ggtitle("By-item by-question raw ratings\nBlack dots indicate by-block means within-item, lines and triangles indicate grand by-block means\nRed dots indicate expected rating")

```

Plot the prior ratings on the two prior questions in the XOr condition in terms of question order:
```{r}
# check whether the ratings in the two prior questions are the same in xor 
d_critical_zScored %>% 
  filter(block != "xor" & block != "some") %>%
  filter(block == class_condition) %>%
  filter(block == "prior", main_type == "xor") %>%
  mutate(priorQ_nr = rep(c(1,2), nrow(.)/2),
         unique_ID = paste(submission_id, ID, sep="_")) -> d_xor_priors

d_xor_priors %>% group_by(priorQ_nr, prior_class) %>%
  summarise(mean = mean(response)) -> d_xor_priors_summary

d_xor_priors %>%
  ggplot(., aes(x = as.factor(priorQ_nr), y = response )) +
  geom_point(size = 2, alpha = 0.6) +
  geom_point(data = d_xor_priors_summary, aes(x = as.factor(priorQ_nr), y = mean), color = "red", size = 3) +
  geom_path(aes(group = "unique_ID"), alpha = 0.6) +
  ylab("Responses to prior questions") +
  xlab("First vs Second prior question for high vs low prior conditions") +
  facet_wrap(~as.factor(prior_class)) + # get ratings from the respective trials only 
  ggtitle("Xor prior question ratings.\nLines indicate the two responses provided by one participant for a specific item.")

```

Plot the prior ratings on the two prior questions in the XOr condition in terms of the specific by-item question (question 1 vs question 2 for a given item, regardless of presentation order):
```{r}
# create labels for the two questions
d_xor_priors %>% group_by(ID, prompt) %>%
  select(ID, prompt) %>% distinct() %>%
  ungroup() %>%
  mutate(
    priorQ_label = rep(c("prior1", "prior2"), 32)
  ) -> d_xor_priors_labels
left_join(d_xor_priors, d_xor_priors_labels, by = c("prompt")) %>%
  group_by(priorQ_label, prior_class) %>%
  summarise(mean = mean(response)) -> d_xor_priors_labels_summary

# add the label and plot
left_join(d_xor_priors, d_xor_priors_labels, by = c("prompt")) %>%
  ggplot(., aes(x = as.factor(priorQ_label), y = response )) +
  geom_point(size = 2, alpha = 0.6) +
  geom_point(data = d_xor_priors_labels_summary, aes(x = as.factor(priorQ_label), y = mean), color = "red", size = 3) +
  geom_path(aes(group = "unique_ID"), alpha = 0.6) +
  ylab("Responses to prior questions") +
  xlab("First vs Second prior question for high vs low prior conditions") +
  facet_wrap(~as.factor(prior_class)) + # get ratings from the respective trials only 
  ggtitle("Xor prior question ratings.\nLines indicate the two responses provided by one participant for a specific item.")
```

Plot inference strength ratings against predictor ratings, by-item, using by-item mean ratings across participants.
```{r, fig.height=10, fig.width=8}
d_critical_predictorsXtarget_summary <-  d_critical_zScored_wide %>% 
  pivot_longer(
    cols = c(relevance, prior, competence),
    names_to = "predictors",
    values_to = "predictor_response"
  ) %>% 
  group_by(title, predictors, main_type) %>%
  summarise(mean_target = mean(target),
            mean_predictor = mean(predictor_response), 
            sd_target = sd(target),
            sd_predictor = sd(predictor_response))  
# get the prior classifications of the vignettes for convenient coloring
d_critical_zScored %>% select(title, class_condition, prior_class) %>%
  distinct() -> d_critical_prior_classes

left_join(d_critical_predictorsXtarget_summary, d_critical_prior_classes, by = c("title", "predictors" = "class_condition")) %>%
  mutate(`Prior categorization` = factor(prior_class, levels = c(0,1),
                                         labels=c("low", "high")),
         Trigger = factor(main_type, levels = c("some", "xor"), 
                            labels = c("Some", "Or")),
         predictors = factor(predictors, levels = c("relevance", "competence", "prior"),
                             labels = c("Relevance", "Competence", "Prior"))) -> df_by_predictor_4_plotting

df_by_predictor_4_plotting %>%
  ggplot(., aes(x = mean_predictor, y = mean_target)) +
  geom_point(aes(color = `Prior categorization`)) +
  ggrepel::geom_text_repel(aes(label=title), size = 3.5) +
  geom_smooth(method = "lm") +
  ylab("Mean inference strength rating") +
  xlab("Mean predictor ratings") + 
  facet_grid(cols=vars(predictors), rows=vars(Trigger)) +
  theme_bw() +
  theme(legend.position = "top",
        legend.text = element_text(size = 10),
        axis.text = element_text(size = 10),
        strip.text.x = element_text(size = 15),
        strip.text.y = element_text(size = 15)) 

ggsave("../../writing/ELM2/byItem-ratings.pdf", width = 9, height = 7)
```

```{r, fig.height=10, fig.width=8}
df_by_predictor_4_plotting %>% 
  group_by(`Prior categorization`, Trigger, predictors) %>%
  summarize(mean = mean(mean_predictor)) -> by_item_predictors_summary

df_by_predictor_4_plotting %>%
  ggplot(., aes(x = `Prior categorization`, y = mean_predictor)) +
  geom_point(aes(color = Trigger, fill = Trigger), size = 3.5, alpha = 0.8,  position = position_jitter(width = 0.2)) +
  #geom_col(data = by_item_predictors_summary, aes(y = mean, x = `Prior categorization`), alpha = 0.3) +
  ylab("Mean predictor strength rating") +
  xlab("Prior categorization of story w.r.t. predictor") + 
  facet_grid(cols=vars(predictors)) + # , rows=vars(main_type)
  theme_bw() +
  theme(legend.position = "top",
        legend.text = element_text(size = 10),
        axis.text = element_text(size = 10),
        strip.text.x = element_text(size = 15),
        strip.text.y = element_text(size = 15)) 
  
ggsave("../../writing/ELM2/byItem-scatter.pdf", width = 9, height = 5)

```

```{r}
df_by_predictor_4_plotting %>%
  ggplot(., aes(x = `Prior categorization`, y = mean_predictor)) +
  geom_point(aes(color = Trigger, fill = Trigger), size = 3.5, alpha = 0.8,  position = position_jitter(width = 0.2)) +
  ggrepel::geom_text_repel(aes(label=title), size = 3.5) +
  ylab("Mean predictor strength rating") +
  xlab("Prior categorization of story w.r.t. predictor") + 
  facet_grid(cols=vars(predictors), rows=vars(Trigger)) +  
  theme_bw() +
  theme(legend.position = "top",
        legend.text = element_text(size = 10),
        axis.text = element_text(size = 10),
        strip.text.x = element_text(size = 15),
        strip.text.y = element_text(size = 15)) 
  
ggsave("../../writing/ELM2/byItem-scatter-wTitles.pdf", width = 9, height = 7)

```

Plot the pairwise correlations of the predictors and the inference rating for "some": 
```{r, message=FALSE, warning=FALSE}
GGally::ggpairs(
  filter(d_critical_zScored_wide, main_type == "some") %>% ungroup() %>%  
    select(prior, competence, relevance, target)
)
```

Plot the pairwise correlations of the predictors and the inference rating for "or": 
```{r, message=FALSE, warning=FALSE}
GGally::ggpairs(
  filter(d_critical_zScored_wide, main_type == "xor") %>% ungroup() %>%  
    select(prior, competence, relevance, target)
)
```

Plot the pairwise correlations of the predictors and the inference rating for "or", by order of presentation of the two prior questions:
```{r, message=FALSE, warning=FALSE}
# correlation plot for "or"
GGally::ggpairs(
  filter(d_critical_zScored_wide_xor_priors_order, main_type == "xor") %>%  ungroup() %>%
    select(prior_1, prior_2, competence, relevance, target)
  )
```

Plot the pairwise correlations of the predictors and the inference rating for "or", by type of prior questions:
```{r, message=FALSE, warning=FALSE}
# correlation plot for "or"
GGally::ggpairs(
  filter(d_critical_zScored_wide_xor_priors_type, main_type == "xor") %>%  ungroup() %>%
    select(priorQ_1, priorQ_2, competence, relevance, target)
  )
```


## Plots for ELM2
Plot predictor ratings against inference strength ratings by trigger:
```{r, echo=FALSE, message=FALSE, warning=FALSE}
# add the prior predictor categorization to wide df
df_by_predictor_4_plotting %>% select(title, predictors, prior_class) %>% 
  mutate(
    predictors = tolower(predictors)
    ) -> df_predictors_prior_classes

d_critical_raw_wPrior <- d_critical_zScored_wide %>% 
  pivot_longer(
    cols = c(relevance, prior, competence),
    names_to = "predictors",
    values_to = "predictor_response"
  ) %>%
  left_join(., df_predictors_prior_classes, by = c("title", "predictors"))

  d_critical_raw_wPrior_4plot <- d_critical_raw_wPrior %>%
    mutate(
      `Prior categorization` = factor(prior_class, levels = c(0,1),
                                         labels=c("low", "high")),
         Trigger = factor(main_type, levels = c("some", "xor"), 
                            labels = c("Some", "Or")),
         predictors = factor(predictors, levels = c("relevance", "competence", "prior"),
                             labels = c("Relevance", "Competence", "Prior"))
    )
```

```{r}
d_critical_raw_wPrior_4plot %>%  
  ggplot(., aes(x = predictor_response, y = target)) +
  geom_point(aes(color = `Prior categorization`), size = 2, alpha = 0.8) +
  geom_smooth(method = "lm") +
  ylab("") +
  facet_grid(cols=vars(predictors), rows=vars(Trigger)) +
  theme_bw() +
  labs(x = "Participants' factor ratings", y = "Participants' inference likelihood ratings", color = "Prior categorization of vignettes") +
  theme(legend.position = "top",
        legend.title = element_text(size = 13),
        legend.text = element_text(size = 13),
        axis.title = element_text(size = 13),
        strip.text.x = element_text(size = 13),
        strip.text.y = element_text(size = 13)) 

#ggsave("../../writing/ELM2/byFactor-byPrior-raw-scatter.pdf", width = 9, height = 7)
```

### Simulated plots for the paper

Below, some simulations are conducted to create plots for the paper. Simulate data similarly to the power analysis, assuming an idealized effect size of 0.2 (somewhat larger than in power analysis, for better visibility). Modify the sampling a little to create low and high condition data.
```{r}
# simulate data from 200 participants with the function from the power analysis
create_fake_data <- function(N = 100, seed = 1) {
  set.seed(seed)
  # optionally changed to 0.25
  beta_rel  <- 0.25
  beta_comp <- 0.25
  beta_pri  <- -0.25
  map_df(
    1:N, 
    function(i) {
      tibble(
        subj    = rep(str_c("subj_",i), times = 8),
        trigger = rep(c('or', 'some'), times= 4), 
        prior = rep(c("low", "high"), each=4),
        rel     = c(rnorm(4, mean=-1), rnorm(4, mean =1)),
       # rel_high = rnorm(4, mean = 0.15),
        comp    = c(rnorm(4, mean=-1), rnorm(4, mean=1)),
       # comp_high=rnorm(4, mean=0.15),
        pri     = c(rnorm(4, mean=1), rnorm(4, mean=-1))
    #    pri_high= rnorm(4, mean=0.15)
      ) %>% 
        group_by(subj) %>% 
        mutate(
          rel     = (rel-mean(rel)) / sd(rel),
          comp    = (comp-mean(comp)) / sd(comp),
          pri     = (pri-mean(pri)) / sd(pri)
        ) %>% 
        ungroup() %>% 
        mutate(
          target  = rnorm(8, beta_rel * rel + beta_comp * comp + beta_pri * pri)
        )
    }
  )
}


fake_data_scatter <- create_fake_data(200, 1234)
```

```{r}
fake_data_long <- fake_data_scatter %>% pivot_longer(
  c(comp, rel, pri), 
               names_to = "Predictors", 
               values_to = "Predictor_response"
) %>%
  mutate(
    prior_categorization = ifelse(Predictor_response <= 0, 0, 1),
    prior_categorization = as.factor(prior_categorization)
  ) %>%
    mutate(
      `Prior categorization` = factor(prior_categorization, levels = c(0,1),
                                         labels=c("low", "high")),
         Trigger = factor(trigger, levels = c("some", "or"), 
                            labels = c("Some", "Or")),
         Predictors = factor(Predictors, levels = c("rel", "comp", "pri"),
                             labels = c("Relevance", "Competence", "Prior"))
    )

# write out for fitting the idealized model in the analysis script
# fake_data_long %>% write_csv("../../data/main/idealized_simulation_long_N200_B025.csv")

fake_data_long %>%
  ggplot(., aes(x = Predictor_response, y = target)) +
  geom_point(aes(color = prior), size = 2, alpha = 0.8) +
  geom_smooth(method = "lm") +
  ylab("") +
  facet_grid(cols=vars(Predictors), rows=vars(Trigger)) +
  theme_bw() +
  labs(x = "Participants' factor ratings", y = "Participants' inference likelihood ratings", color = "Prior categorization of vignettes") +
  theme(legend.position = "top",
        legend.title = element_text(size = 13),
        legend.text = element_text(size = 13),
        axis.title = element_text(size = 13),
        strip.text.x = element_text(size = 13),
        strip.text.y = element_text(size = 13)) 
#ggsave("../../writing/paper/images/idealized_data_N200_b025.pdf",  width = 9, height = 7)
```
